add new directory and folder, impliment fixed and variable window bound calculations
