+++ pandas/tests/window/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-import pandas._libs.custom_window as libwindow_custom
+from pandas.core.window.indexers import BaseIndexer
 
 from pandas import date_range, offsets
 
@@ -55,11 +55,11 @@ def min_periods(request):
 
 @pytest.fixture
 def dummy_custom_indexer():
-    class DummyIndexer(libwindow_custom.BaseIndexer):
-        def __init__(self, index, offset, keys):
-            super().__init__(index, offset, keys)
+    class DummyIndexer(BaseIndexer):
+        def __init__(self, *args, **kwargs):
+            super().__init__()
 
-        def get_window_bounds(self, **kwargs):
+        def get_window_bounds(self):
             pass
 
     idx = date_range("2019", freq="D", periods=3)
